Titanic_challenge_solved_with_decisionTree_aproach
Decision Tree Learning¶
In this document we study the sinking of the RMS Titanic. One of the most famous shipwrecks and one of the most remembered catastrophes in world history.

The sinking of the Titanic led to a huge amount of life loss due to the lack of lifeboats for the passengers and the crew on board.

We explore the different groups that were formed in the shipwreck such as gender, age, classes. In our study we reflect how some of these groups were more likely to survive.

We used data analitics and informatic algorithms to investigate this phenomena and show our results.